Feature/crud

+++ src/main/java/io/github/orodrigobarbosa/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package io.github.orodrigobarbosa.libraryapi.service;
 
+import io.github.orodrigobarbosa.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.orodrigobarbosa.libraryapi.model.Autor;
 import io.github.orodrigobarbosa.libraryapi.repository.AutorRepository;
+import io.github.orodrigobarbosa.libraryapi.repository.LivroRepository;
+import io.github.orodrigobarbosa.libraryapi.validator.AutorValidator;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -9,13 +12,24 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-@AllArgsConstructor
+
 @Service
 public class AutorService {
 
     private AutorRepository autorRepository;
 
+    private AutorValidator autorValidator;
+
+    private LivroRepository livroRepository;
+
+    public AutorService(AutorRepository autorRepository, AutorValidator autorValidator, LivroRepository livroRepository) {
+        this.autorRepository = autorRepository;
+        this.autorValidator = autorValidator;
+        this.livroRepository = livroRepository;
+    }
+
     public Autor salvarAutor(Autor autor) {
+        autorValidator.validar(autor);
         return autorRepository.save(autor);
 
     }
@@ -24,7 +38,7 @@ public class AutorService {
         if (autor.getId() == null) {
             throw new IllegalArgumentException("Para atualizar, se faz necessário que o autor já esteja salvo na base de dados e tenha um id");
         }
-
+        autorValidator.validar(autor);
         return autorRepository.save(autor);
     }
 
@@ -34,6 +48,10 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (autorPossuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Não é permitido deletar. Autor(a) possui livros cadastrados!");
+        }
+
         autorRepository.delete(autor);
     }
 
@@ -57,12 +75,10 @@ public class AutorService {
     }
 
 
+    public boolean autorPossuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
 
-
-
-
-
-
+    }
 
 
 }
